pyproject.toml
[tool.poetry]
name = "llm-lecture"
version = "0.1.0"
description = ""
authors = ["정승현 <tmdgusya@naver.com>"]
readme = "README.md"
packages = [{include = "llm_lecture"}]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^1.0.0"
openai = "^0.27.8"
urllib3 = "1.26.6"
langchain = "^0.0.194"
tiktoken = "^0.4.0"
ipython = "^8.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

